# Optimise embed finder resolution