--- a/wagtail/embeds/apps.py
+++ b/wagtail/embeds/apps.py
@@ -11,5 +11,7 @@ class WagtailEmbedsAppConfig(AppConfig):
     default_auto_field = "django.db.models.AutoField"
 
     def ready(self):
+        from . import signal_handlers  # noqa
+
         # Check configuration on startup
         get_finders()