--- a/wagtail/embeds/embeds.py
+++ b/wagtail/embeds/embeds.py
@@ -9,7 +9,18 @@ from .finders import get_finders
 from .models import Embed
 
 
-def get_embed(url, max_width=None, max_height=None, finder=None):
+def get_finder_for_embed(url, max_width=None, max_height=None):
+    for finder in get_finders():
+        if finder.accept(url):
+            kwargs = {}
+            if accepts_kwarg(finder.find_embed, "max_height"):
+                kwargs["max_height"] = max_height
+            return finder.find_embed(url, max_width=max_width, **kwargs)
+
+    raise EmbedUnsupportedProviderException
+
+
+def get_embed(url, max_width=None, max_height=None, finder=get_finder_for_embed):
     embed_hash = get_embed_hash(url, max_width, max_height)
 
     # Check database
@@ -17,19 +28,6 @@ def get_embed(url, max_width=None, max_height=None, finder=None):
         return Embed.objects.exclude(cache_until__lte=now()).get(hash=embed_hash)
     except Embed.DoesNotExist:
         pass
-
-    # Get/Call finder
-    if not finder:
-
-        def finder(url, max_width=None, max_height=None):
-            for finder in get_finders():
-                if finder.accept(url):
-                    kwargs = {}
-                    if accepts_kwarg(finder.find_embed, "max_height"):
-                        kwargs["max_height"] = max_height
-                    return finder.find_embed(url, max_width=max_width, **kwargs)
-
-            raise EmbedUnsupportedProviderException
 
     embed_dict = finder(url, max_width, max_height)