Currently, finding the embed HTML for a given URL (assuming there's no record in the DB) requires compiling the regex of each oembed provider. This is both CPU and time intensive - and mostly unnecessary.

This change caches the instantiated embed finders. Since regex compilation is done in the constructor, this means it only needs doing once. The finders are initialised during module init (through `apps.py`), which means the cache miss shouldn't hit a web request. This won't interfere with the existing DB cache for the HTML.

I went this route instead of modifying `oembed_finders` to compile them at the module level to also cache class construction, and any externally-added finders or providers.

There was also some duplication in the Facebook / Instagram finders, which I've now removed.

## Before

```
In [2]: %timeit get_finders()
58.4 µs ± 183 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```

## After

```
In [2]: %timeit get_finders()
41.2 ns ± 0.519 ns per loop (mean ± std. dev. of 7 runs, 10,000,000 loops each)
```